refactor: 회원가입 페이지 매직 넘버 상수화
