+++ src/constants/values.ts
@@ -10,4 +10,10 @@ const SECTION_VALUE = {
   DISTANCE_MAX_VALUE: 300,
 };
 
-export { LINE_VALUE, SECTION_VALUE };
+const SIGNUP_VALUE = {
+  AGE_MIN_VALUE: 1,
+  AGE_MAX_VALUE: 200,
+  AGE_MAX_LENGTH: 3,
+};
+
+export { LINE_VALUE, SECTION_VALUE, SIGNUP_VALUE };
